Implementadas validações dos serviços de negócios

+++ src/Business/Services/FornecedorService.cs
@@ -1,5 +1,7 @@
 ﻿using Business.Interfaces;
 using Business.Models;
+using Business.Models.Validations;
+using DevIO.Business.Models.Validations;
 
 namespace Business.Services
 {
@@ -14,20 +16,53 @@ namespace Business.Services
 
 		public async Task Adicionar(Fornecedor fornecedor)
 		{
-			//validar se a entidade é consistente...
+			if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)  || !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) return;
 
-			//validar se ja nao existe outro fornecedor com o mesmo doc
+			if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+			{
+				Notificar("Já existe um fornecedor com este documento informado.");
+				return;
+			}
 
 			await _fornecedorRepository.Adicionar(fornecedor);
 		}
 
 		public async Task Atualizar(Fornecedor fornecedor)
 		{
+			if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)) return;
+
+			if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any())
+			{
+				Notificar("Já existe um fornecedor com este documento informado.");
+				return;
+			}
+
 			await _fornecedorRepository.Atualizar(fornecedor);
 		}
 
 		public async Task Remover(Guid id)
 		{
+			var fornecedor = await _fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+			if (fornecedor == null)
+			{
+				Notificar("Fornecedor não existe!");
+					return;
+			}
+
+			if (fornecedor.Produtos.Any())
+			{
+				Notificar("O fornecedor possui produtos cadastrados!");
+				return;
+			}
+
+			var endereco = await _fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+			if (endereco != null)
+			{ 
+				await _fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+			}
+
 			await _fornecedorRepository.Remover(id);
 		}
